requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0
django-chartjs==2.2.1
django-crispy-forms==1.9.2
django-extensions==3.0.4
django-heroku==0.3.1
django-jazzmin==2.2.6
graphviz==0.14.1
Pillow==7.2.0
psycopg2==2.7.5
psycopg2-binary==2.8.5
pydotplus==2.0.2
pyparsing==2.4.7
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
